event_updater/requirements.txt
functions-framework==3.*
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-cloud-storage>=2.0.0